Move wlog and the waitlistBanner to the /models directory
-Moved the wlog and waitlistBanner files.
- Updated the refrences for those files in apiController, fcToolsController, fleetManagementController and pagesController.
- Removed a few useless imports.
