+++ models/waitlistBanner.js
@@ -1,9 +1,5 @@
-const fs = require('fs');
-const path = require('path');
-const db = require('./dbHandler.js').db.collection('waitlist-banner');
-const ObjectId = require('mongodb').ObjectID;
-const setup = require('./setup.js');
-const log = require('./logger.js')(module);
+const db = require('../dbHandler.js').db.collection('waitlist-banner');
+const log = require('../logger.js')(module);
 
 module.exports = function (setup) {
     var module = {};
